refactor(server): extract /api/content/* into a router module
Tenth route-group extraction. 6 scattered handlers (regenerate-image,
:safeId/latest, :safeId/:contentId, topic-check, :brandSafeId/:contentId,
import) -> src/server/routes/content.js, mounted
`app.use('/api/content', contentRouter)`.

Auth: MIXED -> per-route. GET /:safeId/latest and PATCH
/:brandSafeId/:contentId are open; the rest requireAuth; /import keeps its
dual middleware (requireAuth + requireApiKeyScope('content:import')).

Boundary safety: app.use('/api/content', …) matches on segment
boundaries, so the separate /api/content-library and
/api/content-generator routes (3, different prefixes) are NOT captured —
verified they remain in server.js and the guard still reports them.

Order-sensitive catch-alls preserved: GET /:safeId/latest is registered
before GET /:safeId/:contentId (file order kept), so 'latest' still wins.

Deps all already extracted: pool, anthropic, safeParseLLM,
requireAuth/requireApiKeyScope (auth), stripScaffoldingArtifacts (text),
buildImagePrompt/generateHeroImage (images). No local-helper blockers.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean
- module boot-load (import()) -> resolves
- route-inventory guard -> 213 byte-identical (6 content/ paths moved;
  3 content-library/generator paths still present in server.js)
- vitest -> 93/93

Rollback: revert; verbatim multi-span move.
